fix: remove custom hook
