+++ src/apps/renderer/pages/Settings/Antivirus/index.tsx
@@ -1,5 +1,4 @@
 import { useState } from 'react';
-import { useAntivirus } from '../../../hooks/antivirus/useAntivirus';
 import Button from '../../../components/Button';
 
 interface AntivirusSectionProps {
@@ -7,17 +6,16 @@ interface AntivirusSectionProps {
 }
 
 export default function AntivirusSection({ active }: AntivirusSectionProps) {
-  const { selectItemsToScan, scanSingleDir } = useAntivirus();
   const [selectedItems, setSelectedItems] = useState<{
     path: string;
     itemName: string;
   }>();
 
   const onSelectItemsButtonClicked = async () => {
-    const selectedItems = await selectItemsToScan();
-    if (selectedItems) {
-      setSelectedItems(selectedItems as any);
-    }
+    // const selectedItems = await getPathFromDialog();
+    // if (selectedItems) {
+    //   setSelectedItems(selectedItems as any);
+    // }
   };
 
   return (
